RB-15 Logging with loguru

+++ src/localbase.py
@@ -59,8 +59,11 @@ class LocalBaseWorker:
         return local_name
 
     def remove_database(self, local_name):
-        self.db_session.query(BasesTable).filter_by(local_name=local_name).delete()
-        self.db_session.commit()
+        # Don't execute in local database when local_name is False
+        # In Database/Put except block we pass none
+        if local_name:
+            self.db_session.query(BasesTable).filter_by(local_name=local_name).delete()
+            self.db_session.commit()
 
     def get_database_and_connection(self, local_base_name: str):
         database_obj = (
